feat(theme): overhaul themeing strategy and reduce clutter
remove semantic color overrides as it was too much abstraction.
remove atoms to begin moving away from inline styles.
rethink spacing and sizes to align with UI grid methodology.
remove unnecessary tokens.
