--- src/styles/atoms/index.ts
@@ -1,1 +1,0 @@
-export { atoms } from '@cubeartisan/cubeartisan/styles/atoms/atoms.css';
