+++ src/styles/index.ts
@@ -1,2 +1,1 @@
-export { atoms } from '@cubeartisan/cubeartisan/styles/atoms';
-export { theme, vars } from '@cubeartisan/cubeartisan/styles/themes';
+export { themeClass, vars } from '@cubeartisan/cubeartisan/styles/themes';
