+++ src/styles/themes/colorPaletteType.ts
@@ -13,35 +13,7 @@ export type ColorPaletteType = {
     11: string;
     12: string;
   };
-  primaryA: {
-    1: string;
-    2: string;
-    3: string;
-    4: string;
-    5: string;
-    6: string;
-    7: string;
-    8: string;
-    9: string;
-    10: string;
-    11: string;
-    12: string;
-  };
   neutral: {
-    1: string;
-    2: string;
-    3: string;
-    4: string;
-    5: string;
-    6: string;
-    7: string;
-    8: string;
-    9: string;
-    10: string;
-    11: string;
-    12: string;
-  };
-  neutralA: {
     1: string;
     2: string;
     3: string;
@@ -69,35 +41,7 @@ export type ColorPaletteType = {
     11: string;
     12: string;
   };
-  successA: {
-    1: string;
-    2: string;
-    3: string;
-    4: string;
-    5: string;
-    6: string;
-    7: string;
-    8: string;
-    9: string;
-    10: string;
-    11: string;
-    12: string;
-  };
   info: {
-    1: string;
-    2: string;
-    3: string;
-    4: string;
-    5: string;
-    6: string;
-    7: string;
-    8: string;
-    9: string;
-    10: string;
-    11: string;
-    12: string;
-  };
-  infoA: {
     1: string;
     2: string;
     3: string;
@@ -125,35 +69,7 @@ export type ColorPaletteType = {
     11: string;
     12: string;
   };
-  warningA: {
-    1: string;
-    2: string;
-    3: string;
-    4: string;
-    5: string;
-    6: string;
-    7: string;
-    8: string;
-    9: string;
-    10: string;
-    11: string;
-    12: string;
-  };
   danger: {
-    1: string;
-    2: string;
-    3: string;
-    4: string;
-    5: string;
-    6: string;
-    7: string;
-    8: string;
-    9: string;
-    10: string;
-    11: string;
-    12: string;
-  };
-  dangerA: {
     1: string;
     2: string;
     3: string;
@@ -182,157 +98,3 @@ export type ColorPaletteType = {
     12: string;
   };
 };
-
-// const colorTokens = {
-//   neutral: {
-//     text: {
-//       default: '',
-//       weak: '',
-//       disabled: '',
-//     },
-//     background: {
-//       default: '',
-//       medium: '',
-//       strong: '',
-//       hover: '',
-//       active: '',
-//     },
-//     border: {
-//       default: '',
-//       strong: '',
-//       hover: '',
-//       active: '',
-//     },
-//     icon: {
-//       default: '',
-//       hover: '',
-//     },
-//   },
-//   primary: {
-//     text: {
-//       default: '',
-//       weak: '',
-//     },
-//     background: {
-//       default: '',
-//       weak: '',
-//       strong: '',
-//       hover: '',
-//     },
-//     border: {
-//       default: '',
-//       hover: '',
-//     },
-//     icon: {
-//       default: '',
-//       hover: '',
-//     },
-//   },
-//   success: {
-//     text: {
-//       default: '',
-//       weak: '',
-//     },
-//     background: {
-//       default: '',
-//       weak: '',
-//       hover: '',
-//     },
-//     border: {
-//       default: '',
-//       hover: '',
-//     },
-//     icon: {
-//       default: '',
-//       hover: '',
-//     },
-//   },
-//   info: {
-//     text: {
-//       default: '',
-//       weak: '',
-//     },
-//     background: {
-//       default: '',
-//       weak: '',
-//       hover: '',
-//     },
-//     border: {
-//       default: '',
-//       hover: '',
-//     },
-//     icon: {
-//       default: '',
-//       hover: '',
-//     },
-//   },
-//   warning: {
-//     text: {
-//       default: '',
-//       weak: '',
-//     },
-//     background: {
-//       default: '',
-//       weak: '',
-//       hover: '',
-//     },
-//     border: {
-//       default: '',
-//       hover: '',
-//     },
-//     icon: {
-//       default: '',
-//       hover: '',
-//     },
-//   },
-//   danger: {
-//     text: {
-//       default: '',
-//       weak: '',
-//     },
-//     background: {
-//       default: '',
-//       weak: '',
-//       hover: '',
-//     },
-//     border: {
-//       default: '',
-//       hover: '',
-//     },
-//     icon: {
-//       default: '',
-//       hover: '',
-//     },
-//   },
-//   shadow: {
-//     text: '',
-//     inset: '',
-//     drop: '',
-//   },
-// };
-
-// const colorFramework = {
-//   app: {
-//     bg: '',
-//     text: '',
-//   },
-//   nav: {
-//     bg: '',
-//     text: '',
-//     navButton: {
-//       bg: '',
-//       bgHover: '',
-//       bgActive: '',
-//       text: '',
-//     },
-//   },
-//   hero: {
-//     bg: '',
-//     text: '',
-//     subtext: '',
-//   },
-//   thumbnail: {
-//     insetShadow: '',
-//     text: '',
-//   },
-// };
