+++ src/styles/themes/dark/colorPalette.ts
@@ -1,17 +1,11 @@
 import {
   tomatoDark as danger,
-  tomatoDarkA as dangerA,
   blueDark as info,
-  blueDarkA as infoA,
   mauveDark as neutral,
-  mauveDarkA as neutralA,
   purpleDark as primary,
-  purpleDarkA as primaryA,
   whiteA as shadow,
   greenDark as success,
-  greenDarkA as successA,
   amberDark as warning,
-  amberDarkA as warningA,
 } from '@radix-ui/colors/src';
 
 import type { ColorPaletteType } from '@cubeartisan/cubeartisan/styles/themes/colorPaletteType';
@@ -31,20 +25,6 @@ export const colorPalette: ColorPaletteType = {
     11: neutral.mauve11,
     12: neutral.mauve12,
   },
-  neutralA: {
-    1: neutralA.mauveA1,
-    2: neutralA.mauveA2,
-    3: neutralA.mauveA3,
-    4: neutralA.mauveA4,
-    5: neutralA.mauveA5,
-    6: neutralA.mauveA6,
-    7: neutralA.mauveA7,
-    8: neutralA.mauveA8,
-    9: neutralA.mauveA9,
-    10: neutralA.mauveA10,
-    11: neutralA.mauveA11,
-    12: neutralA.mauveA12,
-  },
   primary: {
     1: primary.purple1,
     2: primary.purple2,
@@ -58,20 +38,6 @@ export const colorPalette: ColorPaletteType = {
     10: primary.purple10,
     11: primary.purple11,
     12: primary.purple12,
-  },
-  primaryA: {
-    1: primaryA.purpleA1,
-    2: primaryA.purpleA2,
-    3: primaryA.purpleA3,
-    4: primaryA.purpleA4,
-    5: primaryA.purpleA5,
-    6: primaryA.purpleA6,
-    7: primaryA.purpleA7,
-    8: primaryA.purpleA8,
-    9: primaryA.purpleA9,
-    10: primaryA.purpleA10,
-    11: primaryA.purpleA11,
-    12: primaryA.purpleA12,
   },
   success: {
     1: success.green1,
@@ -87,20 +53,6 @@ export const colorPalette: ColorPaletteType = {
     11: success.green11,
     12: success.green12,
   },
-  successA: {
-    1: successA.greenA1,
-    2: successA.greenA2,
-    3: successA.greenA3,
-    4: successA.greenA4,
-    5: successA.greenA5,
-    6: successA.greenA6,
-    7: successA.greenA7,
-    8: successA.greenA8,
-    9: successA.greenA9,
-    10: successA.greenA10,
-    11: successA.greenA11,
-    12: successA.greenA12,
-  },
   info: {
     1: info.blue1,
     2: info.blue2,
@@ -114,20 +66,6 @@ export const colorPalette: ColorPaletteType = {
     10: info.blue10,
     11: info.blue11,
     12: info.blue12,
-  },
-  infoA: {
-    1: infoA.blueA1,
-    2: infoA.blueA2,
-    3: infoA.blueA3,
-    4: infoA.blueA4,
-    5: infoA.blueA5,
-    6: infoA.blueA6,
-    7: infoA.blueA7,
-    8: infoA.blueA8,
-    9: infoA.blueA9,
-    10: infoA.blueA10,
-    11: infoA.blueA11,
-    12: infoA.blueA12,
   },
   warning: {
     1: warning.amber1,
@@ -143,20 +81,6 @@ export const colorPalette: ColorPaletteType = {
     11: warning.amber11,
     12: warning.amber12,
   },
-  warningA: {
-    1: warningA.amberA1,
-    2: warningA.amberA2,
-    3: warningA.amberA3,
-    4: warningA.amberA4,
-    5: warningA.amberA5,
-    6: warningA.amberA6,
-    7: warningA.amberA7,
-    8: warningA.amberA8,
-    9: warningA.amberA9,
-    10: warningA.amberA10,
-    11: warningA.amberA11,
-    12: warningA.amberA12,
-  },
   danger: {
     1: danger.tomato1,
     2: danger.tomato2,
@@ -170,20 +94,6 @@ export const colorPalette: ColorPaletteType = {
     10: danger.tomato10,
     11: danger.tomato11,
     12: danger.tomato12,
-  },
-  dangerA: {
-    1: dangerA.tomatoA1,
-    2: dangerA.tomatoA2,
-    3: dangerA.tomatoA3,
-    4: dangerA.tomatoA4,
-    5: dangerA.tomatoA5,
-    6: dangerA.tomatoA6,
-    7: dangerA.tomatoA7,
-    8: dangerA.tomatoA8,
-    9: dangerA.tomatoA9,
-    10: dangerA.tomatoA10,
-    11: dangerA.tomatoA11,
-    12: dangerA.tomatoA12,
   },
   shadow: {
     1: shadow.whiteA1,
@@ -200,32 +110,3 @@ export const colorPalette: ColorPaletteType = {
     12: shadow.whiteA12,
   },
 } as const;
-
-// export const darkTheme = {
-//   neutral: {
-//     text: neutral.mauve1,
-//   },
-//   app: {
-//     bg: neutral.mauve4,
-//     text: primary.purple12,
-//   },
-//   nav: {
-//     bg: secondary.cyan6,
-//     text: secondary.cyan12,
-//     navButton: {
-//       bg: '',
-//       bgHover: '',
-//       bgActive: '',
-//       text: '',
-//     },
-//   },
-//   hero: {
-//     bg: primary.purple6,
-//     text: primary.purple11,
-//     subtext: primary.purple11,
-//   },
-//   thumbnail: {
-//     insetShadow: shadow.whiteA8,
-//     text: neutral.mauve11,
-//   },
-// };
