+++ src/styles/themes/index.ts
@@ -1,2 +1,2 @@
 export { vars } from '@cubeartisan/cubeartisan/styles/themes/vars.css';
-export { theme } from '@cubeartisan/cubeartisan/styles/themes/themeClass.css';
+export { themeClass } from '@cubeartisan/cubeartisan/styles/themes/themeClass.css';
