--- src/styles/themes/light/semanticColorsOverrides.ts
@@ -1,5 +1,0 @@
-import type { DeepPartial } from 'utility-types';
-
-import type { SemanticColors } from '@cubeartisan/cubeartisan/styles/themes/semanticColors';
-
-export const semanticColorOverrides: DeepPartial<SemanticColors> = {};
