+++ src/styles/themes/semanticColors.ts
@@ -359,51 +359,6 @@ export type MakeSemanticColorsProps = {
 };
 
 export const makeSemanticColors = ({ colorPalette, overrides = {} }: MakeSemanticColorsProps): SemanticColors => {
-  /**
-   * (from [Radix Color docs](https://www.radix-ui.com/docs/colorPalette/palette-composition/understanding-the-scale#steps-910-solid-backgrounds))
-   *
-   * Steps 1 and 2 are designed for app backgrounds and subtle component backgrounds.
-   * You can use them interchangeably, depending on the vibe you're going for.
-   *
-   * Appropriate applications include:
-   * - Main app background
-   * - Striped table background
-   * - Code block background
-   * - Card background
-   * - Sidebar background
-   * - Canvas area background
-   *
-   * Steps 3, 4, and 5 are designed for UI component backgrounds.
-   * - Step 3 is for normal states.
-   * - Step 4 is for hover states.
-   * - Step 5 is for pressed or selected states.
-   *
-   * Even if your component has a transparent background in its default state, you should skip Step 3 and use Step 4 for its hover state.
-   *
-   * For call-to-action components that you want to draw particular attention to, you can bump each color one step higher to steps 4, 5, and 6.
-   *
-   * For complex components where you need many subtle colorPalette to communicate different meanings, you can get creative and do something like:
-   * - Step 3 for hovered backgrounds.
-   * - Step 4 for selected / pressed backgrounds.
-   * - Step 5 for "unread" backgrounds.
-   * - Step 6 for hovered "unread" backgrounds.
-   *
-   * Steps 9 and 10 are designed for solid backgrounds.
-   * Step 9 has the highest chroma of all steps in the scale. In other words, it's the purest step,
-   * the step mixed with the least amount of white or black.
-   *
-   * Because 9 is the purest step, it has a wide range of applications:
-   * - Website/App backgrounds
-   * - Website section backgrounds
-   * - Header backgrounds
-   * - Component backgrounds
-   * - Graphics/Logos
-   * - Overlays
-   * - Coloured shadows
-   * - Accent borders
-   *
-   * Step 10 is designed for component hover states, where step 9 is the component's normal state background.
-   */
   const backgroundColor = {
     neutral: {
       neutralSubtle: colorPalette.neutral[1],
