--- src/styles/tokens/index.ts
@@ -1,1 +1,0 @@
-export * from '@cubeartisan/cubeartisan/styles/tokens/tokens';
